style: UI Optimization
